pyproject.toml
[project]
name = "studget"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "dateparser>=1.2.2",
    "fastapi>=0.116.0",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.5.3",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=6.2.0",
    "schemathesis>=4.0.7",
    "sqlmodel[asyncio]>=0.0.24",
    "uvicorn>=0.35.0",
]
